requirements.txt
easyocr==1.7.1
filelock==3.13.1
fsspec==2024.3.1
imageio==2.34.0
Jinja2==3.1.3
lazy_loader==0.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pillow_heif==0.15.0
pyclipper==1.3.0.post5
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.3
six==1.16.0
sympy==1.12
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
tzdata==2024.1
